+++ includes/mybooking-plugin.php
@@ -793,6 +793,7 @@
            mybooking_engine_is_page( $registry->mybooking_rent_plugin_choose_products_page ) ) { // Renting choose product
         $data = array();
         $data['show_taxes_included'] = $registry->mybooking_rent_plugin_show_taxes_included;
+        $data['choose_product_layout'] = get_theme_mod('mybooking_reservation_engine_rent_choose_product_layout', 'list_only');
         mybooking_engine_get_template('mybooking-plugin-choose-product-tmpl.php', $data);
         // If selector in process Wizard, load the micro-templates for the process
         if ($registry->mybooking_rent_plugin_selector_in_process == 'wizard') {